slice

+++ plugins/openflow/controller_api.py
@@ -104,17 +104,19 @@ def create_default_controller(slice_obj, controller_sys):
         raise DbError("数据库异常！")
 
 
-def delete_controller(controller):
+def delete_controller(controller, flag):
     """删除控制器
     """
     if controller:
-        if controller.name == 'user_define' and (not controller.host):
+        if controller.name == 'user_define':
             controller.delete()
         else:
             #先删除虚拟机然后删除controller记录
-#             if controller.host:
-#                 delete_vm_for_controller(controller.host)
-            controller.delete()
+            if flag:
+                if controller.host:
+                    delete_vm_for_controller(controller.host)
+                else:
+                    controller.delete()
 
 
 def slice_change_controller(slice_obj, controller_info):
@@ -144,7 +146,7 @@ def slice_change_controller(slice_obj, controller_info):
             try:
                 print 'c7'
                 if controller:
-                    delete_controller(controller)
+                    delete_controller(controller, True)
                 print 'c8'
                 slice_obj.add_resource(haved_controller)
                 print 'c9'
@@ -154,7 +156,7 @@ def slice_change_controller(slice_obj, controller_info):
             raise
         else:
             try:
-                delete_controller(haved_controller)
+                delete_controller(haved_controller, True)
             except:
                 pass
     else:
